[hba][feat][Tournaments management]

+++ src/Controller/UserController.php
@@ -94,7 +94,6 @@ class UserController extends AbstractController
      * @return JsonResponse // JsonResponse object
      */
     #[Route('players/{id}', name: 'playerById', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN', message: "Only admins can access this route")]
     public function getUserById(User $user, SerializerInterface $serializer): JsonResponse
     {
         $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getPlayers']);
@@ -117,10 +116,14 @@ class UserController extends AbstractController
      * @return JsonResponse // JsonResponse object
      */
     #[Route('players/{id}', name: 'updatePlayer', methods: ['PUT'])]
-    #[IsGranted('ROLE_ADMIN', message: "Only admins can access this route")]
     public function updateUser(Request $request, User $user, SerializerInterface $serializer, ValidatorInterface $validator,
                                UserPasswordHasherInterface $passwordHasher, EntityManagerInterface $em): JsonResponse
     {
+        $currentUser = $this->getUser();
+        if ($currentUser->getId() !== $user->getId() && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            throw new HttpException(Response::HTTP_FORBIDDEN, "You cannot update another user");
+        }
+
         $updatedUser = $serializer->deserialize($request->getContent(), User::class, 'json', ['object_to_populate' => $user]);
 
         $errors = $validator->validate($user);
@@ -148,9 +151,13 @@ class UserController extends AbstractController
      * @return JsonResponse // JsonResponse object
      */
     #[Route('players/{id}', name: 'deletePlayer', methods: ['DELETE'])]
-    #[IsGranted('ROLE_ADMIN', message: "Only admins can access this route")]
     public function deleteUser(User $user, EntityManagerInterface $em): JsonResponse
     {
+        $currentUser = $this->getUser();
+        if ($currentUser->getId() !== $user->getId() && !in_array('ROLE_ADMIN', $currentUser->getRoles())) {
+            throw new HttpException(Response::HTTP_FORBIDDEN, "You cannot delete another user");
+        }
+
         $em->remove($user);
         $em->flush();
         $response = [
